Change the URLs in the API to be more RESTful
