+++ api/src/routes/venue.ts
@@ -3,13 +3,10 @@ import { MysqlError } from "mysql";
 
 const router = Router();
 
-
-router.get("/", (req: Request, res: Response) => {
+router.get("/:id", (req: Request, res: Response) => {
 
 	res.locals.connection
-		.query(req.query.user ?
-			`call getVenuesByUser(${res.locals.connection.escape(req.query.user)})`
-			: "call getAllVenues()",
+		.query(`call getVenueDetails(${res.locals.connection.escape(req.params.id)})`,
 		(error: MysqlError, results: any) => {
 
 			if (error) {
@@ -21,7 +18,7 @@ router.get("/", (req: Request, res: Response) => {
 			} else {
 
 				res.locals.connection.end();
-				res.status(200).json(results[0].map((row: {}) => ({...row})));
+				res.status(200).json(results[0].map((row: {}) => ({...row}))[0]);
 
 			}
